Further tidying and organising of ESRFlvp class
More two more functions into the common file.
Tidy and comment other parts of the class.
